# docs/05-api/openapi.yaml
openapi: 3.1.0
info:
  title: Mira API
  version: 1.0.0
  description: |
    Mira — your closet, your stylist, your mirror.

    Canonical contract for the Mira API. Enumerated values come from
    docs/04-data/taxonomy.md and MUST NOT be extended by implementations.

    Rules that apply to every operation:
      - Responses contain only the authenticated user's data.
      - A resource owned by another user returns 404, never 403.
      - Every creating POST requires an Idempotency-Key header.
      - Image URLs are signed and expiring; clients must honour urlExpiresAt.
servers:
  - url: https://api.mira.app/v1
    description: Production
  - url: https://api.staging.mira.app/v1
    description: Staging
  - url: http://localhost:4000/v1
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /garments:
    get:
      tags: [garments]
      summary: List garments
      parameters:
        - { in: query, name: category, schema: { type: array, items: { type: string } }, style: form, explode: true }
        - { in: query, name: subcategory, schema: { type: array, items: { type: string } }, style: form, explode: true }
        - { in: query, name: brand_id, schema: { type: array, items: { type: string, format: uuid } }, style: form, explode: true }
        - { in: query, name: color, schema: { type: array, items: { type: string } }, style: form, explode: true }
        - { in: query, name: size, schema: { type: array, items: { type: string } }, style: form, explode: true }
        - { in: query, name: season, schema: { type: array, items: { type: string } }, style: form, explode: true }
        - { in: query, name: occasion, schema: { type: array, items: { type: string } }, style: form, explode: true }
        - { in: query, name: material, schema: { type: array, items: { type: string } }, style: form, explode: true }
        - { in: query, name: style_tag, schema: { type: array, items: { type: string } }, style: form, explode: true }
        - { in: query, name: retailer, schema: { type: array, items: { type: string } }, style: form, explode: true }
        - { in: query, name: status, schema: { type: array, items: { $ref: '#/components/schemas/GarmentStatus' } }, style: form, explode: true }
        - { in: query, name: favorite, schema: { type: boolean } }
        - { in: query, name: tags_attached, schema: { type: boolean } }
        - { in: query, name: never_worn, schema: { type: boolean } }
        - { in: query, name: not_worn_since_days, schema: { type: integer, minimum: 1 } }
        - { in: query, name: purchased_after, schema: { type: string, format: date } }
        - { in: query, name: purchased_before, schema: { type: string, format: date } }
        - { in: query, name: price_min, schema: { type: number } }
        - { in: query, name: price_max, schema: { type: number } }
        - in: query
          name: sort
          schema:
            type: string
            enum: [recent, recently_worn, never_worn, brand, color, price_desc, price_asc]
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of garments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GarmentPage'
    post:
      tags: [garments]
      summary: Create a garment
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GarmentCreate'
      responses:
        '200':
          description: >-
            Merged into an existing garment. Nothing was created, so the client
            must not add a second tile for a piece it is already showing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Garment'
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Garment'
        # duplicate_unresolved — this may already be in the closet. `details`
        # names each candidate; call check-duplicate for the imagery the sheet
        # needs (D-024).
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /garments/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [garments]
      summary: Get a garment
      responses:
        '200':
          description: The garment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Garment'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [garments]
      summary: Update a garment
      description: source.type is immutable and MUST be rejected if present.
      parameters:
        - in: header
          name: If-Match
          schema: { type: string }
          description: The garment's updatedAt value, for optimistic concurrency.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GarmentUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Garment'
        '409':
          $ref: '#/components/responses/Conflict'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [garments]
      summary: Remove a garment (soft delete)
      responses:
        '204':
          description: Removed
  /garments/{id}/favorite:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [garments]
      summary: Set favourite state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [favorite]
              properties:
                favorite: { type: boolean }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Garment'
  /garments/{id}/status:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [garments]
      summary: Set garment status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/GarmentStatus' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Garment'
...
  /garments/check-duplicate:
    post:
      tags: [ingestion]
      summary: Check whether a garment about to be created already exists
      description: Called by every ingestion path before creation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DuplicateCheckRequest'
      responses:
        '200':
          description: Duplicate candidates, most likely first
          content:
            application/json:
              schema:
                type: object
                properties:
                  candidates:
                    type: array
                    items:
                      $ref: '#/components/schemas/DuplicateCandidate'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 40 }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, maxLength: 128 }
  responses:
...
    NotFound:
      description: Not found, or not the caller's resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Concurrency or idempotency conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  issue: { type: string }
            request_id: { type: string }
            retry_after: { type: integer, nullable: true }
...
    Money:
      type: object
      properties:
        amount: { type: number }
        currency: { type: string, minLength: 3, maxLength: 3 }
    GarmentStatus:
      type: string
      enum: [active, laundry, unavailable, lent_out, returned, sold, donated, lost, archived]
    SourceType:
      type: string
      enum: [manual, camera, photo_library, tag_scan, barcode, receipt, email, retailer_integration, product_url, order_screenshot]
...
    ImageKind:
      type: string
      enum: [canonical, original, cleaned, front, back, side, detail, retailer]
    GarmentImage:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: '#/components/schemas/ImageKind' }
        url: { type: string, format: uri }
        # 400px WebP, for the closet grid. Null until image.process has run, or
        # if derivative generation failed — clients fall back to `url` rather
        # than showing nothing (docs/06-ai/image-processing.md §8).
        thumb_url: { type: string, format: uri, nullable: true }
        # 1080px WebP, for garment and look detail.
        medium_url: { type: string, format: uri, nullable: true }
        url_expires_at: { type: string, format: date-time }
        width: { type: integer, nullable: true }
        height: { type: integer, nullable: true }
        blurhash: { type: string, nullable: true }
        is_canonical: { type: boolean }
        position: { type: integer }
    Brand:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        logo_url: { type: string, nullable: true }
    Garment:
      type: object
      required: [id, category, status, source]
      properties:
        id: { type: string, format: uuid }
        closet_id: { type: string, format: uuid }
        name: { type: string, nullable: true }
        brand: { $ref: '#/components/schemas/Brand' }
        brand_raw: { type: string, nullable: true }
        category: { type: string, description: taxonomy §1 }
        subcategory: { type: string, nullable: true }
        primary_color: { type: string, nullable: true }
        secondary_colors: { type: array, items: { type: string } }
        pattern: { type: string, nullable: true }
        materials: { type: array, items: { type: string } }
        size:
          type: object
          properties:
            raw: { type: string, nullable: true }
            normalized: { type: string, nullable: true }
            system: { type: string, nullable: true }
        fit: { type: string, nullable: true }
        season: { type: array, items: { type: string } }
        occasion: { type: array, items: { type: string } }
        style_tags: { type: array, items: { type: string } }
        purchase:
          type: object
          properties:
            date: { type: string, format: date, nullable: true }
            price: { $ref: '#/components/schemas/Money' }
            retailer: { type: string, nullable: true }
        identifiers:
          type: object
          properties:
            sku: { type: string, nullable: true }
            barcode: { type: string, nullable: true }
            product_url: { type: string, nullable: true }
        source:
          type: object
          description: Immutable provenance.
          properties:
            type: { $ref: '#/components/schemas/SourceType' }
            reference: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/GarmentStatus' }
        favorite: { type: boolean }
        tags_attached: { type: boolean, nullable: true }
        notes: { type: string, nullable: true }
        wear:
          type: object
          properties:
            count: { type: integer }
            last_worn_at: { type: string, format: date-time, nullable: true }
            cost_per_wear: { $ref: '#/components/schemas/Money' }
        images:
          type: array
          items: { $ref: '#/components/schemas/GarmentImage' }
        canonical_image: { $ref: '#/components/schemas/GarmentImage' }
        analysis_state:
          type: string
          enum: [pending, analyzing, complete, failed, skipped]
        confidence:
          type: object
          additionalProperties: { type: number, minimum: 0, maximum: 1 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    GarmentCreate:
      type: object
      required: [category]
      properties:
        name: { type: string, nullable: true }
        brand_raw: { type: string, nullable: true }
        category: { type: string }
        subcategory: { type: string, nullable: true }
        primary_color: { type: string, nullable: true }
        secondary_colors: { type: array, items: { type: string } }
        pattern: { type: string, nullable: true }
        materials: { type: array, items: { type: string } }
        size_raw: { type: string, nullable: true }
        fit: { type: string, nullable: true }
        season: { type: array, items: { type: string } }
        occasion: { type: array, items: { type: string } }
        style_tags: { type: array, items: { type: string } }
        purchase_date: { type: string, format: date, nullable: true }
        purchase_price: { type: number, nullable: true }
        currency: { type: string, nullable: true }
        retailer: { type: string, nullable: true }
        sku: { type: string, nullable: true }
        barcode: { type: string, nullable: true }
        product_url: { type: string, nullable: true }
        source_type: { $ref: '#/components/schemas/SourceType' }
        source_reference: { type: string, nullable: true }
        image_upload_keys: { type: array, items: { type: string } }
        tags_attached: { type: boolean, nullable: true }
        notes: { type: string, nullable: true }
        # The user's answer to the duplicate sheet (duplicate-detection.md §4).
        # Omit it and a create that would duplicate is refused with 409
        # duplicate_unresolved (CAP-5).
        #
        # This was a bare enum of [none, owns_two, different], which could not
        # express §4: it had no way to name WHICH garment — ambiguous the moment
        # there is more than one candidate — and no `same_item`, so the merge
        # that §5 spends a section defining had no way to be requested.
        duplicate_resolution:
          $ref: '#/components/schemas/DuplicateResolution'
    GarmentUpdate:
      allOf:
        - $ref: '#/components/schemas/GarmentCreate'
        - type: object
          description: source_type is immutable; sending it results in 422.
    GarmentPage:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Garment' }
        next_cursor: { type: string, nullable: true }
        total: { type: integer, nullable: true }
...
    DuplicateCheckRequest:
      $ref: '#/components/schemas/GarmentCreate'
    DuplicateCandidate:
      type: object
      properties:
        existing_garment: { $ref: '#/components/schemas/Garment' }
        score: { type: number, minimum: 0, maximum: 1 }
        # What to do, so a client never has to know the thresholds in §3.
        # `ask` and `ask_softly` show the sheet; `note` is recorded quietly and
        # surfaced later in wardrobe insights.
        band:
          type: string
          enum: [ask, ask_softly, note]
        # One line for the sheet: "Same brand and a very similar name" (§4).
        summary: { type: string }
        # One name per row of duplicate-detection.md §2. `visual_similarity`
        # arrives with Phase 5 embeddings; the rest are computed today.
        signals:
          type: array
          items:
            type: string
            enum:
              - barcode
              - sku_retailer
              - product_url
              - order_line
              - image_hash
              - visual_similarity
              - brand_name
              - category_color_size_brand
              - purchase_window
    DuplicateResolution:
      type: object
      required: [garment_id, relation]
      properties:
        garment_id: { type: string, format: uuid }
        relation:
          type: string
          enum: [same_item, owns_two, different]
...